Don’t hash production server files
Long story short: in addition to the app/ server build folder, we require the `app/` source folder to generate the route manifest on the server. If we deploy the app source folder vercel wants to compile the TS for us, so instead we only ship the built app folder and branch on an environment variable in remix.config.js in starter-vercel to make the route manifest.

Because of this, we can’t have hash names in the server entries because the routes/loader naming convention breaks.

No need for hashes on the server build anyway.