Inline loader/action in route modules
This is a major change that adds support for inlining `loader` and
`action` functions directly in route modules and removes support for
putting them in modules in a separate `data` directory.

Packages that are used only in `loader` and/or `action` functions are
automatically pruned from the browser build using Rollup's built-in
tree-shaking capabilities.