send just pathname in manifest requests
Passing the whole URL overcomplicates the upcoming link export feature, requiring the server code to pass the protocol/host/port all the way down to the code that generates the link href, only to have the manifest request toss everything out but the pathname.

Turns out, pathname was the only argument we needed to create the manifest request, too.

Pathname is all we knew, pathname is all we used. We constructed a full URL in-between for no reason.